fragment-validator/requirements.txt
pandas==2.1.4
boto3==1.34.0
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary>=2.9.0
